% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plots.R
\name{Bin_data}
\alias{Bin_data}
\title{Function that splits the data into bins and computes the average in each bin}
\usage{
Bin_data(Y, X, c = NULL, binsize = 0.2)
}
\arguments{
\item{Y}{\itemize{
\item is the outcome data
}}

\item{X}{\itemize{
\item is the score data
}}

\item{c}{\itemize{
\item specifies the cutoff point, set to NULL if the binning of the data should be independent from c
}}

\item{binsize}{\itemize{
\item length of the interval that is one bin
}}
}
\value{
list with elements Y_bin: list of average values of Y in each bin, X_bin: list of middle points for each bin
}
\description{
Function that splits the data into bins and computes the average in each bin
}
