% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plots.R
\name{CONT_outcome_function_sample}
\alias{CONT_outcome_function_sample}
\title{Function that evaluates the continuous outcome function in a domain x, given the coefficients}
\usage{
CONT_outcome_function_sample(coef_s, x, d_x, s_x, mu_y, sd_y)
}
\arguments{
\item{coef_s}{\itemize{
\item list with the function coefficients
}}

\item{x}{\itemize{
\item points at which the function is evaluated
}}

\item{d_x}{\itemize{
\item shifting parameter that was used to normalize x
}}

\item{s_x}{\itemize{
\item scaling parameter that was used to normalize x
}}

\item{mu_y}{\itemize{
\item shifting parameter that was used to normalize y
}}

\item{sd_y}{\itemize{
\item scaling parameter that was used to normalize y
}}
}
\value{
list with the values of the function
}
\description{
Function that evaluates the continuous outcome function in a domain x, given the coefficients
}
