% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_model.R
\name{Initial_CONT_CONT}
\alias{Initial_CONT_CONT}
\title{function that samples initial values for fuzzy LoTTA model with a ontinuous prior and continuous outcomes}
\usage{
Initial_CONT_CONT(x, t, y, C_start, clb, cub, lb, ubr, ubl, s, jlb = 0.2)
}
\arguments{
\item{x}{\itemize{
\item score data
}}

\item{t}{\itemize{
\item treatment data
}}

\item{y}{\itemize{
\item outcome data
}}

\item{C_start}{\itemize{
\item posterior samples of cutoff location
obtained through "cutoff_initial_CONT.txt"
}}

\item{clb}{\itemize{
\item left end of an interval on which the prior is supported
}}

\item{cub}{\itemize{
\item right end of an interval on which the prior is supported
}}

\item{lb}{\itemize{
\item minimum window size
}}

\item{ubr}{\itemize{
\item maximum value of the window's right boundary point
}}

\item{ubl}{\itemize{
\item minimum value of the window's left boundary point
}}

\item{s}{\itemize{
\item seed
}}

\item{jlb}{\itemize{
\item minimum jump size
}}
}
\value{
list with initial parameters values for LoTTA model with continuous score
and continuous outcomes, and .RNG.seed value
}
\description{
function that samples initial values for fuzzy LoTTA model with a ontinuous prior and continuous outcomes
}
