% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_model.R
\name{Initial_FUZZY_BIN}
\alias{Initial_FUZZY_BIN}
\title{function that samples initial values for fuzzy LoTTA model with a known cutoff and binary outcomes}
\usage{
Initial_FUZZY_BIN(X, T, Y, c, lb, ubr, ubl, s, jlb = 0.2)
}
\arguments{
\item{X}{\itemize{
\item score data
}}

\item{T}{\itemize{
\item treatment data
}}

\item{Y}{\itemize{
\item outcome data
}}

\item{c}{\itemize{
\item cutoff location
}}

\item{lb}{\itemize{
\item minimum window size
}}

\item{ubr}{\itemize{
\item maximum value of the window's right boundary point
}}

\item{ubl}{\itemize{
\item minimum value of the window's left boundary point
}}

\item{s}{\itemize{
\item seed
}}

\item{jlb}{\itemize{
\item minimum jump size
}}
}
\value{
list with initial parameters values for fuzzy LoTTA model with a known cutoff and
binary outcomes, and .RNG.seed value
}
\description{
function that samples initial values for fuzzy LoTTA model with a known cutoff and binary outcomes
}
