% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_model.R
\name{Initial_SHARP_BIN}
\alias{Initial_SHARP_BIN}
\title{function that samples initial values for sharp LoTTA model with binary outcomes}
\usage{
Initial_SHARP_BIN(X, Y, c, lb, ubr, ubl, s)
}
\arguments{
\item{X}{\itemize{
\item score data
}}

\item{Y}{\itemize{
\item outcome data
}}

\item{c}{\itemize{
\item cutoff location
}}

\item{lb}{\itemize{
\item minimum window size
}}

\item{ubr}{\itemize{
\item maximum value of the window's right boundary point
}}

\item{ubl}{\itemize{
\item minimum value of the window's left boundary point
}}

\item{s}{\itemize{
\item seed
}}
}
\value{
list with initial parameters values for sharp LoTTA model with
binary outcomes, and .RNG.seed value
}
\description{
function that samples initial values for sharp LoTTA model with binary outcomes
}
