% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_model.R
\name{Initial_treatment_CONT}
\alias{Initial_treatment_CONT}
\title{function that samples initial values for the treatment model with a continuous prior}
\usage{
Initial_treatment_CONT(X, T, C_start, clb, cub, lb, ubr, ubl, s, jlb = 0.2)
}
\arguments{
\item{X}{\itemize{
\item score data
}}

\item{T}{\itemize{
\item treatment data
}}

\item{C_start}{\itemize{
\item posterior samples of cutoff location
obtained through "cutoff_initial_CONT.txt"
}}

\item{clb}{\itemize{
\item left end of an interval on which the prior is supported
}}

\item{cub}{\itemize{
\item right end of an interval on which the prior is supported
}}

\item{lb}{\itemize{
\item minimum window size
}}

\item{ubr}{\itemize{
\item maximum value of the window's right boundary point
}}

\item{ubl}{\itemize{
\item minimum value of the window's left boundary point
}}

\item{s}{\itemize{
\item seed
}}

\item{jlb}{\itemize{
\item minimum jump size
}}
}
\value{
list with initial parameters values for LoTTA model with continuous score
and .RNG.seed value
}
\description{
function that samples initial values for the treatment model with a continuous prior
}
