% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_model.R
\name{invlogit}
\alias{invlogit}
\title{inverse logit function}
\usage{
invlogit(x)
}
\arguments{
\item{x}{\itemize{
\item score data
}}
}
\value{
value of inverse logit function at x
}
\description{
inverse logit function
}
