% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_cont_Y}
\alias{normalize_cont_Y}
\title{normalize continuous outcome function}
\usage{
normalize_cont_Y(Y, X, trimmed = NULL, normalize = TRUE)
}
\arguments{
\item{Y}{\itemize{
\item outcome data
}}

\item{X}{\itemize{
\item score data
}}

\item{trimmed}{\itemize{
\item NULL by default, provide vector of values to trim the data
}}

\item{normalize}{\itemize{
\item whether data is to be normalized
}}
}
\value{
list with Y - normalized outcome, mu - shift parameter, sd - scaling parameter
}
\description{
normalize continuous outcome function
}
