% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_dis_X}
\alias{normalize_dis_X}
\title{normalize discrete score function}
\usage{
normalize_dis_X(X, trimmed = NULL, normalize = TRUE)
}
\arguments{
\item{X}{score data}

\item{trimmed}{\itemize{
\item NULL by default, provide vector of values to trim the data
}}

\item{normalize}{\itemize{
\item TRUE by default, whether data is to be normalized
}}
}
\value{
list with X - normalized score, s - parameter used to scale the score
}
\description{
normalize discrete score function
}
