% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_model.R
\name{optimal_k_bin}
\alias{optimal_k_bin}
\title{function that searches for initial parameters of binary outcome function to initiate the sampler}
\usage{
optimal_k_bin(X, Y, c, lb, ubr, ubl)
}
\arguments{
\item{X}{\itemize{
\item score data
}}

\item{Y}{\itemize{
\item outcome data
}}

\item{c}{\itemize{
\item cutoff
}}

\item{lb}{\itemize{
\item minimum window size
}}

\item{ubr}{\itemize{
\item maximum value of the window's right boundary point
}}

\item{ubl}{\itemize{
\item minimum value of the window's left boundary point
}}
}
\value{
\itemize{
\item a list with kl and kr and cubic functions parameters
}
}
\description{
function that searches for initial parameters of binary outcome function to initiate the sampler
}
