% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_model.R
\name{read_prior}
\alias{read_prior}
\title{function that checks the type of a prior and whether it is correct}
\usage{
read_prior(prior, minx, maxx, ubl, ubr, lb)
}
\arguments{
\item{prior}{\itemize{
\item list with prior parameters or a cutoff value
}}

\item{minx}{\itemize{
\item minimum value of the score
}}

\item{maxx}{\itemize{
\item maximum value of the score
}}

\item{ubl}{\itemize{
\item minimum value of the window's left boundary point
}}

\item{ubr}{\itemize{
\item maximum value of the window's right boundary point
}}

\item{lb}{\itemize{
\item minimum window size
}}
}
\value{
list with type of prior and prior parameters
}
\description{
function that checks the type of a prior and whether it is correct
}
