% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{trim_dis_y}
\alias{trim_dis_y}
\title{Binary outcomes for trimmed score}
\usage{
trim_dis_y(y, x, trimmed = NULL)
}
\arguments{
\item{y}{\itemize{
\item outcome data
}}

\item{x}{\itemize{
\item score data
}}

\item{trimmed}{\itemize{
\item NULL by default, provide vector of values to trim the data
}}
}
\value{
list with y - outcomes for trimmed score
}
\description{
Binary outcomes for trimmed score
}
