% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_cont_X}
\alias{normalize_cont_X}
\title{normalize continuous score function}
\usage{
normalize_cont_X(X, trimmed = NULL, normalize = TRUE)
}
\arguments{
\item{X}{\itemize{
\item score data
}}

\item{trimmed}{\itemize{
\item NULL by default, provide vector of score values to trim the data
}}

\item{normalize}{\itemize{
\item whether data is to be normalized
}}
}
\value{
list with X - normalized score, d - parameter used to shift the score
}
\description{
normalize continuous score function
}
