% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{trim_dis_Y}
\alias{trim_dis_Y}
\title{Binary outcomes for trimmed score}
\usage{
trim_dis_Y(Y, X, trimmed = NULL)
}
\arguments{
\item{Y}{\itemize{
\item outcome data
}}

\item{X}{\itemize{
\item score data
}}

\item{trimmed}{\itemize{
\item NULL by default, provide vector of values to trim the data
}}
}
\value{
list with Y - outcomes for trimmed score
}
\description{
Binary outcomes for trimmed score
}
